backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.7.4
python-multipart==0.0.9
google-generativeai==0.8.5
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.2.0
supabase==2.3.4
python-dotenv==1.0.1
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
tiktoken>=0.7.0
